// CAC/src/app/model/catalogo/itemFiltroRequest.model.ts
export class ItemFiltroRequest {

    constructor() {
    }
    pageNumber: number = 0;
    maxRows: number = 0;
    id: string = "";
    catalogoCoberturaId: string = "";
    itemId: string = "";
    itemDescripcion: string = "";
    itemActivo: string = "";
    itemOrden: string = "";
    itemGlosa: string = "";   
   
}